# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Roisfaozi_queue-base_internal_modules_queue_model.ForwardQueueRequest:
    properties:
      destination_counter_id:
        type: string
      destination_service_id:
        type: string
    required:
    - destination_service_id
    type: object
  github_com_Roisfaozi_queue-base_internal_modules_queue_model.RegisterQueueRequest:
    properties:
      patient_id:
        type: string
      patient_name:
        maxLength: 255
        minLength: 2
        type: string
      service_id:
        type: string
    required:
    - patient_name
    - service_id
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.PageMetadata:
    properties:
      limit:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
      total_item:
        type: integer
      total_page:
        type: integer
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper:
    properties:
      data: {}
      paging:
        $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.PageMetadata'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a clean and modular boilerplate for Go REST APIs with RBAC,
    Audit Logs, and WebSockets.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Clean Boilerplate API
  version: "1.0"
paths:
...
  /branches/{id}/counters/{counter_id}/queue-journeys:
    get:
      description: Returns active queue journeys filtered by branch and counter under
        active tenant scope.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Counter ID
        in: path
        name: counter_id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Queue date
        in: query
        name: queue_date
        type: string
      - description: Journey status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: List active journeys by branch and counter
      tags:
      - queues
  /branches/{id}/queue-stats:
    get:
      description: Returns queue statistics for a branch under active tenant scope.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get queue stats by branch
      tags:
      - queues
  /branches/{id}/services/{service_id}/queue-journeys:
    get:
      description: Returns active queue journeys filtered by branch and service under
        active tenant scope.
      parameters:
      - description: Branch ID
        in: path
        name: id
        required: true
        type: string
      - description: Service ID
        in: path
        name: service_id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Queue date
        in: query
        name: queue_date
        type: string
      - description: Journey status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: List active journeys by branch and service
      tags:
      - queues
...
  /queues:
    get:
      description: Returns queues under active tenant and branch scope.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Queue status
        in: query
        name: status
        type: string
      - description: Queue date
        in: query
        name: queue_date
        type: string
      - description: Service ID
        in: query
        name: service_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: List queues
      tags:
      - queues
    post:
      consumes:
      - application/json
      description: Creates one queue master row and first queue journey under tenant
        and branch scope.
      parameters:
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Register Queue Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_queue_model.RegisterQueueRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Register queue
      tags:
      - queues
  /queues/{id}:
    get:
      description: Returns queue details under active tenant scope.
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get queue by ID
      tags:
      - queues
  /queues/{id}/forward:
    post:
      consumes:
      - application/json
      description: Forwards queue by appending queue journey and preserving master
        queue row.
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      - description: Forward Queue Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_Roisfaozi_queue-base_internal_modules_queue_model.ForwardQueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Forward queue
      tags:
      - queues
  /queues/{id}/visit-journeys:
    get:
      description: Returns readable visit history for a queue under active tenant
        scope.
      parameters:
      - description: Queue ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Organization-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerSuccessResponseWrapper'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Roisfaozi_queue-base_pkg_response.SwaggerErrorResponseWrapper'
      security:
      - BearerAuth: []
      summary: Get visit journeys by queue
      tags:
      - queues
...
securityDefinitions:
  BearerAuth:
    description: '"Type ''Bearer '' followed by a space and the access token."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"